+++ src/concepts/speechBubbles/index.ts
@@ -1,4 +1,4 @@
-// import { VideoGui } from "backdop/utils/babylonjs/VideoGui";
+// import { VideoGui } from "backdopdev/utils/babylonjs/VideoGui";
 //
 import { CSSProperties } from "react";
 import { forEach } from "shutils/dist/loops";